% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{phenotype_filter_wt}
\alias{phenotype_filter_wt}
\title{Filter out rows where the phenotype is "wt"}
\usage{
phenotype_filter_wt(n)
}
\arguments{
\item{n}{A data frame containing a column named 'phenotype'.}
}
\value{
A data frame with rows where the 'phenotype' column is not "wt".
}
\description{
This function takes a data frame as input and returns a new data frame
with rows where the 'phenotype' column is not equal to "wt".
}
\examples{
# Assuming you have a data frame named data
# run phenotype_filter_wt(data) and save it to a name you want to use
}
