% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_order.R
\name{set_phenotype_order}
\alias{set_phenotype_order}
\title{Set the order of phenotype levels in the overextension data.}
\usage{
set_phenotype_order(n, phenotype_column, ...)
}
\arguments{
\item{n}{A data frame containing a column named 'phenotype'.}

\item{phenotype_column}{The name of the column that contains the phenotypes, in quotation marks}

\item{...}{list the name of each phenotype in the order you want them, with the first in the list
being the top of the stacked bar graph. Each should be in quotation marks with a comma separating them.}
}
\value{
The input data frame with the 'phenotype' column converted to an \strong{ordered factor}
with levels set according to the list provided for ...
}
\description{
This function reorders the levels of the 'phenotype' factor column in a data frame.
You can either provide a specific order for the levels or let the function
use the unique phenotype names in the order they appear in the data.
}
\examples{

# Example usage with a custom order of levels
# ordered_data_pheno <- set_phenotype_order(example_data_pheno, "phenotype", "severe", "mod")
# levels(ordered_data_pheno$phenotype)
}
