% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_gaps_bargraph.R
\name{tau_gap_bargraph}
\alias{tau_gap_bargraph}
\title{Create a bar graph of tau gap scoring.}
\usage{
tau_gap_bargraph(
  data,
  x_label = "Genotype",
  y_label = "# of Gaps",
  plot_title = "your title here"
)
}
\arguments{
\item{data}{A data frame containing the data for the bar graph.
It should have columns named:
\itemize{
\item{\strong{KO}}: A factor or character column representing the KO genotype
\item{\strong{gaps}}: A numeric column containing the number of gaps.
\item{\strong{day}}: A numeric or character column representing the day,
\item{\strong{error}}: A numeric column containing the error values for
the error bars.
}}

\item{x_label}{what you want the x axis to be labeled as}

\item{y_label}{what you want to y axis to be labeled as}

\item{plot_title}{what you want the title of the plot to be.}
}
\value{
A ggplot object representing the bar graph.
}
\description{
This function takes a data frame as input and generates a bar graph
visualizing tau gap scores, grouped by KO type and day, with error bars.
}
\examples{
#tau_gap_plot <- tau_gap_bargraph(tau_data, "KO", "Gaps", "Scoring Gaps in the Motor Cord")
#tau_gap_plot

}
