% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_gaps_df.R
\docType{data}
\name{tau_gaps}
\alias{tau_gaps}
\title{Tau Gap Scoring Data}
\format{
A data frame with each row representing an individual observation. The columns include:
\describe{
\item{KO}{The genotype of the animal (knockout strain name).}
\item{day}{The experimental day on which the data was collected (either "1" or "6").}
\item{gaps}{The mean number of gaps observed for the given KO and day.}
\item{error}{The standard error associated with the gap count.}
}
}
\source{
Data derived from the \code{tau_gaps.xlsx} file located in the \verb{inst/extdata/} directory
of the Borg-kit package.
}
\usage{
tau_gaps
}
\description{
This dataset contains results from tau gap scoring assays for various knockout (KO) genotypes
across two time points (Day 1 and Day 6). It is used to evaluate gap formation phenotypes
in C. elegans under different genetic conditions.
}
\keyword{datasets}
